Bug 1321168 - Create separate EBS volume on b-2008 and y-2008
EBS volumes initialized as part of an AMI (such as the root volume)
are extremely slow on first block access. They are inappropriate
for doing heavy work on. Instead, you need to create a separate,
empty EBS volume and do I/O intensive work there.

In this commit, we establish a new EBS volume on the b-2008 and
y-2008 machine configs.

The existing Mount-EphemeralDisks PowerShell function for
processing extra volumes didn't exactly inspire confidence that
it would handle the case of extra volumes not being ephemeral
volumes. So, I've rewritten the function. It now does slightly more
robust validation. Although the function is still severely limited
in functionality.

While I was here, I also removed "instance_dev" from the device_map
because it is only used when calling `resize2fs` (which doesn't happen
on Windows) and I reduced the size of the root volume because it
doesn't need to be so large since most work happens on the "scratch"
EBS volume.