mozilla-releng.net route53 resources generated from ...
configuration in mozilla-releng/services

  $ ./please tools terraform-route53-config

Problem we are trying to solve/avoid is that with growing number of
services it is hard to keep track what is deployed to production and/or
staging.

For this reason we added new subcomand to ./please command that
generates route53 configuration, since what is deployed where is already
stored aloside with with code.